// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TransmissionStatus = {
  PENDING: 'PENDING',
  RECEIVED: 'RECEIVED',
  READ: 'READ',
  ARCHIVED: 'ARCHIVED'
} as const

export type TransmissionStatus = (typeof TransmissionStatus)[keyof typeof TransmissionStatus]


export const TypeDocument = {
  DNI: 'DNI',
  RUC: 'RUC',
  PASAPORTE: 'PASAPORTE'
} as const

export type TypeDocument = (typeof TypeDocument)[keyof typeof TypeDocument]


export const InscriptionType = {
  STUDENTS: 'STUDENTS',
  BOOKINGS: 'BOOKINGS'
} as const

export type InscriptionType = (typeof InscriptionType)[keyof typeof InscriptionType]


export const TypeDebt = {
  BOOKING: 'BOOKING',
  INSCRIPTION: 'INSCRIPTION',
  MONTH: 'MONTH',
  PER_SESSION: 'PER_SESSION'
} as const

export type TypeDebt = (typeof TypeDebt)[keyof typeof TypeDebt]


export const PayMethod = {
  CASH: 'CASH',
  BANK: 'BANK',
  QR: 'QR'
} as const

export type PayMethod = (typeof PayMethod)[keyof typeof PayMethod]


export const DayOfWeek = {
  MONDAY: 'MONDAY',
  TUESDAY: 'TUESDAY',
  WEDNESDAY: 'WEDNESDAY',
  THURSDAY: 'THURSDAY',
  FRIDAY: 'FRIDAY',
  SATURDAY: 'SATURDAY',
  SUNDAY: 'SUNDAY'
} as const

export type DayOfWeek = (typeof DayOfWeek)[keyof typeof DayOfWeek]


export const Gender = {
  MASCULINO: 'MASCULINO',
  FEMENINO: 'FEMENINO'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const EducationLevel = {
  PRIMARIA: 'PRIMARIA',
  SECUNDARIA: 'SECUNDARIA'
} as const

export type EducationLevel = (typeof EducationLevel)[keyof typeof EducationLevel]


export const AcademicStatus = {
  ACTIVO: 'ACTIVO',
  INACTIVO: 'INACTIVO',
  EGRESADO: 'EGRESADO',
  TITULADO: 'TITULADO',
  MAESTRIA: 'MAESTRIA'
} as const

export type AcademicStatus = (typeof AcademicStatus)[keyof typeof AcademicStatus]


export const TypeAction = {
  manage: 'manage',
  create: 'create',
  read: 'read',
  update: 'update',
  delete: 'delete'
} as const

export type TypeAction = (typeof TypeAction)[keyof typeof TypeAction]


export const AttendanceStatus = {
  PENDING: 'PENDING',
  PRESENT: 'PRESENT',
  ABSENT: 'ABSENT',
  LATE: 'LATE',
  JUSTIFIED: 'JUSTIFIED'
} as const

export type AttendanceStatus = (typeof AttendanceStatus)[keyof typeof AttendanceStatus]